Add SwiftyJSON library. Temporarily patch the upload() issue. Add documention.
SwiftyJSON is now included in the source code. It is used to parse responses from the Watson Speech to Text service's REST api.

The upload() function was failing because of a time-out error. As a result, no response was returned from the server. Sending a GET request immediately prior to upload()'s POST request often fixes the problem. It doesn't always seem to work, though, so there may be a race condition. The REST API also does not seem to require any communication prior to the POST, so the GET should be extraneous and eventually refactored in favor of a permanent solution.

Extensive documentation was added to the WatsonSpeechToText.swift library, including links to helpful external documentation, an example call/response, a list of design decisions/questions, and a list of todo items.