Created a negative test case for authentication
This test does not pass yet because we need to change the response from the token granting server instead of returning a String with either "Watson rejected token" or simply the Watson token. Instead, it will return a JSON object with a status parameter and a field for the returned token.