Check whether token is refreshing before checking if its nil
When a new AuthenticationStrategy is instantiated we expect it to have a nil token. If many requests are made while the token is nil (i.e. before the refreshToken operation completes), then the WatsonGateway would refresh the token for each request.

This commit switches to order to first check whether the token is refreshing and then check if it is nil.

By switching the order we can ensure that only the first request with a nil token causes the refreshToken method to be called. Subsequent requests will be cached until the token has been refreshed.