Add WebSocketManager
WebSocketManager abstracts a WebSocket. The WebSocketManager accepts strings and raw data that should be sent over a WebSocket, and internally connects to the socket if it is not already.

This primarily keeps us from having to manage the connection state in the SpeechToText service. Instead, it can be managed as a lower-level abstraction.