Set `isConnecting` immediately
Previously we waited until calling `socket.connect()` to set `isConnecting` to true. But if the token needs to be refreshed before calling `socket.connect()`, then `isConnecting` might not be set until after the response from a REST request.

In this scenario, if another attempt is made to connect to the websocket while waiting to retrieve the token, then `isConnecting` would remain false, `retries` may exceed the maximum, and an error might be erroneously produced.