Modify SpeechToText.createTranscriptionObject's return type
The createTranscriptionOutput function now returns a tuple where the first element is the AVCaptureAudioDataOutput object that was returned before, but the second element is a function that stops streaming to Speech to Text.

This addresses and oversight that we discovered while testing the createTranscriptionOutput function in an example application -- we were unable to manually close the connection to Speech to Text before.