Track the state of the SpeechToTextService
This ensures that we pause and unpause the operations queue at the right times (i.e. unpause the queue when in the "Listening" state; pause the queue when disconnected).

This allows us to insert a "pause" and then a "disconnect" operation into the queue, such that the socket is disconnected only after all queued recognition requests have succeeded.