Add user agent to WatsonGateway and NetworkUtils
This commit adds a user-agent HTTP header to requests made with WatsonGateway and NetworkUtils. The user agent begins with the string: "watson-developer-cloud-ios-0.2.0" and also contains information about the executable, bundle, version, and OS.

Note that this commit does NOT add the user-agent to the multipart form uploads, which are executed directly by Alamofire. It also does NOT add the user-agent to the socket used by Speech to Text.