Add error handling to RestKit
Adds an additional parameter for a closure that specifies how to convert NSData to an NSError object. This is useful for many of the Watson APIs, where error information is returned as a JSON object.

By parsing the object with the given closure, we can create an NSError that propagates back to the user with important error information.