Visual Recognition v3 (#253)
* Add foundation for VisualRecognition

This commit includes the models and networking endpoints for the Visual Recognition V2 service. It still requires tests and documentation.

* Adjust VisualRecognition model names and comments

* Add initial VisualRecognition test framework

* Add more verbose error handling to RestSerialization

This commit explicitly catches errors produced by the Freddy JSON parsing library. These errors are propagated to the caller with more detailed information than before.

* Update optional properties for VisualRecognition models

Testing showed these properties to be optional. We have adjusted the model to express this.

* Fix minor bugs in VisualRecognition

* Add tests for VisualRecognition

Note that these tests currently refer to a "VisualRecognitionAPIKey" key in the Credentials.plist file. When our test key is approved for use with the Visual Recognition service, this key should be updated to "AlchemyAPIKey" instead.