Move models out of service extensions
We previously placed all models within an extension of their respective service class. This was needed to create a make-shift namespace for each service and avoid conflicting model names (e.g. `Profile` in multiple services). However, the target for each service now creates a proper namespace, and these extensions are no longer necessary.

This commit moves the models outside of the extensions, and corrects the indentation of the affected models.