Add profanity filter to TranscriptionSettings
As @mitulgolakiya noted in #257, the `filterProfanity` boolean was never sent to the server. As a result, the profanity filter always defaulted to `true`, even if the `filterProfanity` property was `false`.

This commit sends the profanity filter property to the server in the start message that declares the settings.

This commit also re-arranges the order of the properties in the `toJSON` function so that they are consistent with their declaration above. This helps to verify that all properties are being sent to the server.

(I checked, and `filterProfanity` was the only option that was not being sent to the server.)