Fix bugs in the `rest-kit` branch prior to release (#259)
* Update deprecated error handling in RestSerialization

The `Error.errorWithCode()` function was deprecated in Alamofire.

* Remove magic array indices from Speech to Text models

* Handle failure with JSON.serializeString() in SpeechToTextAudioStreamer

Instead of silently failing, this constructs an NSError and executes the `failure()` function.

* Rename ClassWithExamples to Class

Note that `Class` conflicts with a previous model. This commit merges both models.

* Rename FaceImages->ImagesWithFaces, FaceImage->ImageWithFaces

These names better describe what their models represent.

* Rename WordImages->ImagesWithWords, WordImage->ImageWithWords

These names better describe what their models represent.

* Rename VisualRecognitionV3 model files

Update models to reflect the name of their defined structs.

* Remove TODO that has already been completed

This TODO was completed in an earlier commit. This removes the TODO from the code base.

* Update access modifier of Class.name

* Remove VisualRecognition helper functions that duplicate functionality

* Create targets for each service

This commit creates release and test targets for each service. This change has several benefits:

1. It resolves an issue with identically-named files. These files are now associated with separate targets, and so they no longer conflict.

2. It makes it easier for clients to choose the service version they would like to use.

3. It reduces the amount of code that must be copied into client applications. Instead of copying and importing the entire WatsonDeveloperCloud target, clients can now selectively choose the services they would like to use.

* Change json.array(...) to json.arrayOf(..., type: ...)

json.arrayOf(..., type: ...) is more expressive than constructing an array and mapping each JSON value to the given type.

* Move models out of service extensions

We previously placed all models within an extension of their respective service class. This was needed to create a make-shift namespace for each service and avoid conflicting model names (e.g. `Profile` in multiple services). However, the target for each service now creates a proper namespace, and these extensions are no longer necessary.

This commit moves the models outside of the extensions, and corrects the indentation of the affected models.

* Update copyright date to 2016

As per the team's decision earlier, this commit updates the copyright date on all rest-kit files to 2016.

* Adjust formatting of type declarations

Ensure that all type declarations (except for those in the AlchemyLanguage target--which will soon be replaced) are formatted as follows: `variableName: Type` rather than `variableName:Type` or `variableName :Type`.

* Remove unnecessary string interpolation

Variables that are already strings do not need to be interpolated as `"\(stringVariable)"`, instead they can be used directly as `stringVariable`.

* Remove print statement in VisualRecognition

* Add copyright header

* Remove version name from service classes

We previously used the name of each service to distinguish between different versions (e.g. `public class ToneAnalyzerV1`). Since the project has been reorganized to use targets, this is no longer necessary. Instead, the version is already identified in the import statement (e.g. `import ToneAnalyzerV1`).

* Add profanity filter to TranscriptionSettings

As @mitulgolakiya noted in #257, the `filterProfanity` boolean was never sent to the server. As a result, the profanity filter always defaulted to `true`, even if the `filterProfanity` property was `false`.

This commit sends the profanity filter property to the server in the start message that declares the settings.

This commit also re-arranges the order of the properties in the `toJSON` function so that they are consistent with their declaration above. This helps to verify that all properties are being sent to the server.

(I checked, and `filterProfanity` was the only option that was not being sent to the server.)

* Update ToneAnalysis to v3

This commit updates ToneAnalysis for consistency with the v3 API.

* Update version date in VisualRecognitionTests

* Update model comments for `init` and `toJSON` functions

* Add scheme to build all services and tests

This scheme is not shared in order to prevent it from being built by Carthage.

* Update service init and dataToError function comments

* Update error name to reflect package name

Now with separate targets for each service, it makes more sense for the error domain to reflect the name of the target instead of `WatsonDeveloperCloud`.