Modify relative links in readme
Since the readme is included in the documentation generated by Jazzy, we want the links in the readme to be valid from the Jazzy documentation.

Unfortunately, links with a relative path (e.g. Documentation/Quickstart.md) are not updated by Jazzy to point to the proper location. As a result, we replaced relative links with absolute links.