Add properties to match JSON response
These properties may be considered "internal" to the service, but since they are returned in the service's response, we believe they should be captured in the iOS SDK.

If this becomes a support problem, we should communicate to the Watson team to modify the response. Then we can modify the SDK accordingly.