Image handling (#325)
* Use NSData instead of NSURL for easier Image Handling

Using NSURL invokes some trouble with Image Handling. If you choose a Photo from your Camera Roll you get an UIImage or a referencing URL to the Asset. The Initializer NSData(contentsOfURL: ...) does not handle Asset URLs correctly. It seems to me, that NSData reduces a lot of pains for the developers. They can choose what they want to use: Loading from Assets and convert to NSData or transforming UIImage to an UIImagePNGRepresentation.

* Adjust Tests to use NSData instead of NSURL

* Adjust documentation to new API