Add the Relationship Extraction service. (#324)
* Created RelationshipExtraction target

* Added a few models

* Committing to save changes

* Adding final model

* Adding the relationship extraction method

* Updated models so that the JSON will serialize properly.

* Added some comments to the models

* Finished commenting the models

* Added final comments and a few negative unit tests

* Update the json.decode, as well as some comments and variable names

* Removed internal variables from models. Updated unit test to be more thorough when checking the model.

* Add properties to match JSON response

These properties may be considered "internal" to the service, but since they are returned in the service's response, we believe they should be captured in the iOS SDK.

If this becomes a support problem, we should communicate to the Watson team to modify the response. Then we can modify the SDK accordingly.