Improve generateDocumentation.sh
This commit improves the generateDocumentation.sh script that uses Jazzy to generate the documentation for the gh-pages branch of the repository.

In particular, this makes it easier to add/remove services from a list of targets. It also generates the index.html file with links to each service's documentation. All generated documentation is in a gh-pages/ folder at the root of the repository and can be immediately committed to the gh-pages/ branch after execution.