Update tests (#384)
* Add tests to RelationshipExtractionV1Beta scheme

* Speed up Dialog tests

This commit speeds up the Dialog test suite by using a single dialog application instead of creating a new dialog application for each test.

Note that some test errors were introduced and will be fixed in an upcoming commit.

* Download dialog files to downloads directory

This commit modifies the `getDialogFile` method to ensure that dialog files are downloaded to the user’s downloads directory with a unique name and proper extension.

* Update the getDialogFile test function

- Verify that the file exists at the given path
- Verify that the extension of the downloaded file is correct
- Remove the downloaded file to avoid collisions with subsequent tests

* Update getDialogFile error handling

* Update status code for getDialogFile negative test

* Avoid time-out in DialogTests

* Update VisualRecognitionTests to use an already trained classifier

* Adjust VisualRecognitionTests timeout

* Update Dialog tests

* Use separate credentials for ConversationExperimental

* Update Dialog tests

* Update LanguageTranslator tests

* Update TextToSpeech customization models

The service is returning time in milliseconds since the Unix epoch instead of seconds. The resulting values are too large for a Swift `Int` and cannot be parsed by Freddy. These fields are optional, so that they will be populated after the bug has been fixed.

* Update TextToSpeech tests