Revert changes to SolrCluster
I misinterpreted the response and caused an error with changes in an earlier commit. Since `solrClusterSize` is returned as a String, it can be the empty string (“”) which does not map to an Int. Therefore, forcing a value with the guard statements caused the tests to fail.

This commit reverts the earlier changes and once again makes the type of `solrClusterSize` an optional Int.