Improve Speech to Text implementation
This commit improves the Speech to Text implementation.

The SpeechToText class now provides simpler functions for transcribing audio files, transcribing audio data, and starting/stopping streaming transcriptions with the microphone.

The new SpeechToTextSession class provides power-users with additional information, customizability, and control over the WebSocket connection to the service.

The implementation now accesses the microphone with audio queues instead of an AVCaptureSession, which makes for cleaner code and more flexibility.

The new SpeechToTextEncoder class is used to encode PCM microphone audio to Opus before streaming it to the Speech to Text service. This is expected to dramatically improve performance.