Improve recognizeMicrophone
This commit renames the `recognize` function in SpeechToText that starts the microphone to `recognizeMicrophone`. User testing showed that the previous naming caused confusion, because it wasn’t clear what was being recognized (the other `recognize` functions take a parameter for a file or audio data, but this does not).

This commit also stores the state of the microphone session internally. Instead of returning an object with a `finish()` method, the user will invoke `stopRecognizeMicrophone()` in the SpeechToText class. This was also a point of confusion discovered in user testing—it was strange for recognizeMicrophone to return a value when the other recognie functions do not, and it seemed unusual to need to use that object to stop the microphone instead of calling a function in the SpeechToText class.