Generalize bestTranscript
This commit changes bestTranscript from a computed property to a function. As a function, it now accepts a `combine` argument allowing users to specify how multiple transcripts should be combined.

For example, should they be separated by just “ “? Should they be terminated with periods? Should their first letters be capitalized?