Update SpeechRecognitionResults
- Add comments
- Revert bestTranscript to a computed property

Writing a comment for the bestTranscript function revealed its complexity. Since is the primary entry-point for the results, I think we want it to be simple. So I’ve reverted it to a computed property. Users can still traverse the results array if they want to combine the transcripts in a different way.