Improve Speech to Text implementation (#440)
* Improve Speech to Text implementation

This commit improves the Speech to Text implementation.

The SpeechToText class now provides simpler functions for transcribing audio files, transcribing audio data, and starting/stopping streaming transcriptions with the microphone.

The new SpeechToTextSession class provides power-users with additional information, customizability, and control over the WebSocket connection to the service.

The implementation now accesses the microphone with audio queues instead of an AVCaptureSession, which makes for cleaner code and more flexibility.

The new SpeechToTextEncoder class is used to encode PCM microphone audio to Opus before streaming it to the Speech to Text service. This is expected to dramatically improve performance.

* Review Speech to Text implementation

This commit is the result of a thorough review of the new Speech to Text implementation. It includes minor updates, bug fixes, comments, and tests.

* Add default values for model and learningOptOut

* Rename file parameter to audio

* Rename file parameter to audio

* Reset results when starting a recognition request

* Add raw microphone data callback

* Support compressed audio with raw data callback

This commit changes the raw data callback to be invoked after microphone data is compressed to Opus (if applicable). This is consistent with the speech-ios-sdk.

* Add power callback

* Add power timer to current run loop

* Invoke power callback every 0.025s instead of 0.125s

* Carefully start/stop recording

This commit ensures that the microphone is not already being recorded when executing startRecording, and that it was recording before executing stopRecording. It also invalidates the power timer and disposes of the queue when recording is stopped.

* Remove recorder print statement

* Remove state from SpeechToTextSession

The state does not seem necessary for the user. So we will publish our initial release without the state property.

* Remove onListening callback from SpeechToTextSession

This callback does not seem necessary for the user. So we will publish our initial release without it.

* Check record status in SpeechToTextSession

Request record permission and avoid starting the audio queues if permission was not granted.

* Update model names to match API reference

* Add bestTranscript property to SpeechRecognitionResult

This computed property makes it easier for clients to access the most confident transcription.

* Update SpeechToTextTests to reflect model name changes

* Add initial readme update for Speech to Text service

* Add SpeechRecognitionResults wrapper for results array

This model will manage the results array that is constructed from a Speech to Text recognition request. It also exposes a computed property, bestTranscript, that makes it easier to construct a transcript from the results.

* Construct recorder and encoder only when using microphone

By constructing the recorder and encoder in the startMicrophone function, we can also ensure that the ogg/opus encoder’s state is reset and ready for another stream.

* Add SpeechRecognitionResult.swift to project

* Update readme for Speech to Text

* Only reset encoder not recorder

* Update SpeechToTextTests to reflect model changes

* Rename callbacks

onPower -> onPowerData
onAudio -> onMicrophoneData

* Improve recognizeMicrophone

This commit renames the `recognize` function in SpeechToText that starts the microphone to `recognizeMicrophone`. User testing showed that the previous naming caused confusion, because it wasn’t clear what was being recognized (the other `recognize` functions take a parameter for a file or audio data, but this does not).

This commit also stores the state of the microphone session internally. Instead of returning an object with a `finish()` method, the user will invoke `stopRecognizeMicrophone()` in the SpeechToText class. This was also a point of confusion discovered in user testing—it was strange for recognizeMicrophone to return a value when the other recognie functions do not, and it seemed unusual to need to use that object to stop the microphone instead of calling a function in the SpeechToText class.

* Add support for retrieving models

* Add test for retrieving models

* Generalize bestTranscript

This commit changes bestTranscript from a computed property to a function. As a function, it now accepts a `combine` argument allowing users to specify how multiple transcripts should be combined.

For example, should they be separated by just “ “? Should they be terminated with periods? Should their first letters be capitalized?

* Empty queue when disconnecting

* Don’t queue messages when disconnected

This commit adds a new state, `Connecting` to capture when the session is connecting to the service. It also adds guard statements to writeStart, writeAudio, writeStop, and writeNop to avoid queing any messages when disconnected.

* Update SpeechRecognitionResults

- Add comments
- Revert bestTranscript to a computed property

Writing a comment for the bestTranscript function revealed its complexity. Since is the primary entry-point for the results, I think we want it to be simple. So I’ve reverted it to a computed property. Users can still traverse the results array if they want to combine the transcripts in a different way.

* Improve comments for SpeechToText and SpeechToTextSession

* Modify os_types.h to avoid including a non-modular header

* Remove model and learningOptOut parameters from RecognitionSettings

These parameters are still configurable, but are set elsewhere.

* Remove Ogg/Opus imports

The ogg and opus header files are now imported in the umbrella header rather than specified as a module map. Although this exposes the libogg and libopus functions to users of the framework, it prevents the additional installation step of specifying the module search paths.

* Update onMicrophoneData comment

* Update readme for SpeechToText

* Remove outdated Speech to Text example

* Import opus_multistream in umbrella header

* Configure audio session

Physical devices require an AVAudioSession to be defined and configured.

* Update SpeechToTextTests

* Add config_types.h to umbrella header

* Remove config_types.h from the framework

It seems to cause a warning in app projects that use the SpeechToText framework. We can either remove it from the framework or add it to the umbrella header as an import.

* Manage AVAudioSession in SpeechToTextSession

* Remove print statements

* Modify AVAudioSession configuration