Move service url declarations
This is just a minor style change. I personally prefer to list any public properties before the private properties. Since the public properties are commented, I also prefer to add a space between them (otherwise it can sometimes be unclear whether the comment is associated with the property above or below it).