Update service urls (#444)
* First pass at updating service URLs to public vars

* Moved some comments, since we removed parameters from the init statements.

* Added some comments, made the socket for STT initialize lazily

* Move service url declarations

This is just a minor style change. I personally prefer to list any public properties before the private properties. Since the public properties are commented, I also prefer to add a space between them (otherwise it can sometimes be unclear whether the comment is associated with the property above or below it).

* Update init style

* Add more descriptive comments for AlchemyLanguage init

* Add more descriptive comments for AlchemyDataNews init

* Remove default values from SpeechToTextSocket init

These values should be set in SpeechToTextSession. Removing these default values helps us to ensure that we always provide values for these arguments.

* Add comments to SpeechToText

* Remove trailing whitespace character in SpeechToTextSession

Was this really necessary? Not one bit. But once I saw it, I couldn’t unsee it!