Update Visual Recognition
This commit updates the Visual Recognition service to align with the current API Explorer page. Note that there is a bug with uploading images (most likely because of the multipart/form data payload). As a result, the functions that upload images have been commented out in preparation for the upcoming release. We expect to tackle this bug very soon and publish an updated minor release that adds these functions back into the API.