Ensure data payload is not empty
We have occassionally seen an error message related to insufficient data in the audio stream. We suspect it is caused by writing zero bytes to the websocket, which is interpreted by the service as the end of the stream. This commit guards against sending 0 bytes when streaming from the microphone.

Note that users can still use the SpeechToTextSession class and accidentally send a 0-byte Data value. In that case, we want the error to propagate back to the user, which is why we didn’t add these guard statements to the SpeechToTextSocket.writeAudio(audio:) function.