Remove decodeIfPresent for additional properties
We will change the type of additionalProperties from [String: JSON]? to [String: JSON]. Rather than representing the lack of additionalProperties as `nil`, we will represent it as an empty dictionary.

Doing so simplifies the use of additionalProperties for our users. Particularly when setting an additional property, they can avoid having to check for `nil` and creating a new dictionary themselves.