Add sleep command before deleting classifiers
It takes a few seconds for image data to propagate to Object Storage. If the data hasn’t propagated yet, then it is left in limbo when the classifier is deleted. The service instance can be slowed down quite a lot as a result. Adding sleep commands helps to prevent the data from getting out of sync.