Sleep to force the unix time stamp to increment
The access token seems to be a hash of the API key and unix time stamp. If the test were to start and finish executing during the same unix second, then the access token would be the same before and after refreshing.

Sleeping for a second ensures that the unix time increments. As a result, the access token is always different before and after refreshing.