ci(travis): require master branch stages to be a non-PR non-fork
we have recently enabled running PR builds to test PRs against the master branch. However, there is
some undesired behavior with travis using the Jobs conditional syntax where "branch = master" is
True on a PR (since they run on the base branch). This change makes it so that we only run the
deploy stages if we are actually pushing to the master branch (after a PR has been merged).