ci(https git): clone repo using HTTPS instead of SSH
SSH cloning requires a manual SSH RSA fingerprint verification. If you look in the build history of
the master branch, you can see that "publish documentation" always times out waiting for an
interactive verification. Trying to automate this approval would be insecure, so therefore since we
are just cloning a public repo, we should use HTTPS.